+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/craftengine/CraftEngineIntegrator.java
@@ -1,8 +1,11 @@
 package org.betonquest.betonquest.compatibility.craftengine;
 
 import org.betonquest.betonquest.api.BetonQuestApi;
+import org.betonquest.betonquest.api.quest.QuestTypeRegistries;
 import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.Integrator;
+import org.betonquest.betonquest.compatibility.craftengine.action.CraftEngineSetBlockActionFactory;
+import org.betonquest.betonquest.compatibility.craftengine.action.CraftEngineSetFurnitureActionFactory;
 import org.betonquest.betonquest.compatibility.craftengine.item.CraftEngineItemFactory;
 import org.betonquest.betonquest.compatibility.craftengine.item.CraftEngineQuestItemSerializer;
 import org.betonquest.betonquest.item.ItemRegistry;
@@ -21,6 +24,11 @@ public class CraftEngineIntegrator implements Integrator {
 
     @Override
     public void hook(final BetonQuestApi api) throws HookException {
+        final QuestTypeRegistries questTypeRegistries = api.getQuestRegistries();
+
+        questTypeRegistries.action().register("craftEngineBlockAt", new CraftEngineSetBlockActionFactory());
+        questTypeRegistries.action().register("craftEngineFurnitureAt", new CraftEngineSetFurnitureActionFactory());
+
         final ItemRegistry itemRegistry = api.getFeatureRegistries().item();
         itemRegistry.register("craftEngine", new CraftEngineItemFactory());
         itemRegistry.registerSerializer("craftEngine", new CraftEngineQuestItemSerializer());