+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/nexo/NexoIntegrator.java
@@ -1,8 +1,11 @@
 package org.betonquest.betonquest.compatibility.nexo;
 
 import org.betonquest.betonquest.api.BetonQuestApi;
+import org.betonquest.betonquest.api.quest.QuestTypeRegistries;
 import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.Integrator;
+import org.betonquest.betonquest.compatibility.nexo.action.NexoSetBlockActionFactory;
+import org.betonquest.betonquest.compatibility.nexo.action.NexoSetFurnitureActionFactory;
 import org.betonquest.betonquest.compatibility.nexo.item.NexoItemFactory;
 import org.betonquest.betonquest.compatibility.nexo.item.NexoQuestItemSerializer;
 import org.betonquest.betonquest.item.ItemRegistry;
@@ -21,6 +24,11 @@ public class NexoIntegrator implements Integrator {
 
     @Override
     public void hook(final BetonQuestApi api) throws HookException {
+        final QuestTypeRegistries questTypeRegistries = api.getQuestRegistries();
+
+        questTypeRegistries.action().register("nexoBlockAt", new NexoSetBlockActionFactory());
+        questTypeRegistries.action().register("nexoFurnitureAt", new NexoSetFurnitureActionFactory());
+
         final ItemRegistry itemRegistry = api.getFeatureRegistries().item();
         itemRegistry.register("nexo", new NexoItemFactory());
         itemRegistry.registerSerializer("nexo", new NexoQuestItemSerializer());