+++ CHANGELOG.md
@@ -123,6 +123,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `hologram` `top:` line to customizable format
 - Variables to Placeholders. This change won't harm any user scripts, it's only a conceptual change
 - Events to Actions.
+- `nexo` item type
+- `craftEngine` item type
+- `itemsAdder` item type
+- `nexoBlockAt`, `nexoFurnitureAt` action type
+- `craftEngineBlockAt`, `craftEngineFurnitureAt` action type
+- `itemsAdderBlockAt`, `itemsAdderFurnitureAt`, `itemsAdderPlayAnimation` action type
 ### Deprecated
 ### Removed
 - undocumented prefix feature in conversation
@@ -219,9 +225,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
     - `traincartsride` condition to check if the player is riding a specific named train
 - `scoretag` event and condition
 - `action` objective now supports configuring the hand used for the interaction, preventing multiple objective completions at the same time
-- `nexo` item type
-- `craftEngine` item type
-- `itemsAdder` item type
 ### Changed
 - `spawn` event now only spawn mobs and no other entities
 - ProSkillAPI rename to Fabled